Build instead of using images
The images on docker isn't really up-to date, and not having the latest ebot changes.
Might be me that's just a noob. But spent way too long finding a solution that worked for me.
This is just a proposed solution. There should maybe be a image, and build options - but not that good at the docker-compose yet.
Or maybe we can setup a auto-build on dockerhub, for when the original ebot github gets an update.